+++ src/libstd/user/functions/io/getdents.c
@@ -6,7 +6,7 @@
 #include "common/print.h"
 #include "user/common/syscalls.h"
 
-uint64_t getdents(fd_t fd, dirent_t* buffer, uint64_t count)
+size_t getdents(fd_t fd, dirent_t* buffer, uint64_t count)
 {
     uint64_t result = _syscall_getdents(fd, buffer, count);
     if (result == ERR)
