+++ src/libstd/user/functions/io/ioctl.c
@@ -6,7 +6,7 @@
 #include "common/print.h"
 #include "user/common/syscalls.h"
 
-uint64_t ioctl(fd_t fd, uint64_t request, void* argp, uint64_t size)
+uint64_t ioctl(fd_t fd, uint64_t request, void* argp, size_t size)
 {
     uint64_t result = _syscall_ioctl(fd, request, argp, size);
     if (result == ERR)
