+++ src/libstd/user/functions/io/seek.c
@@ -6,7 +6,7 @@
 #include "common/print.h"
 #include "user/common/syscalls.h"
 
-uint64_t seek(fd_t fd, int64_t offset, seek_origin_t origin)
+size_t seek(fd_t fd, ssize_t offset, seek_origin_t origin)
 {
     uint64_t result = _syscall_seek(fd, offset, origin);
     if (result == ERR)
