+++ src/libstd/user/functions/proc/mprotect.c
@@ -3,7 +3,7 @@
 
 #include "user/common/syscalls.h"
 
-void* mprotect(void* address, uint64_t length, prot_t prot)
+void* mprotect(void* address, size_t length, prot_t prot)
 {
     void* result = _syscall_mprotect(address, length, prot);
     if (result == NULL)
